requirements.txt
azure-identity
azure-mgmt-resource
azure-mgmt-subscription
python-dotenv
yagmail
